backend/requirements.txt
fastapi
uvicorn
ollama
pydantic
python-dotenv
asyncpg
pymupdf
python-multipart
langchain-text-splitters
httpx
torch 
torchvision 
torchaudio 
marker
marker-pdf[full]

sqlalchemy 
psycopg2-binary 
asyncpg
pgvector
langchain-community 
langchain-google-ai
langchain-google-genai